requirements.txt
apache-airflow==2.9.3
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
pytest==8.1.1
pytest-mock==3.14.0
